isolate toggling of datepicker
