+++ Baby's Diary.xcodeproj/xcuserdata/tomirisrahymzan.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -14,10 +14,10 @@
             filePath = "Baby&apos;s Diary/Baby/AddBabyViewController.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "56"
-            endingLineNumber = "56"
-            landmarkName = "AddBabyViewController"
-            landmarkType = "21">
+            startingLineNumber = "39"
+            endingLineNumber = "39"
+            landmarkName = "dateSelection(_:didSelectDate:)"
+            landmarkType = "7">
          </BreakpointContent>
       </BreakpointProxy>
    </Breakpoints>
